+++ packages/implementations/react/src/atomic/Stepper/src/components/Step/index.ts
@@ -1,0 +1,5 @@
+import { Step } from "./Step";
+
+export { Step } from "./Step";
+
+export default Step;