+++ packages/implementations/react/src/atomic/Stepper/src/components/Step/step.types.ts
@@ -1,0 +1,10 @@
+import { HTMLAttributes, ReactNode } from "react";
+
+export interface StepProperties {
+  title?: string;
+  description?: string;
+  done?: boolean;
+  active?: boolean;
+  children: ReactNode;
+}
+export type StepProps = StepProperties & HTMLAttributes<HTMLDivElement>;