+++ packages/implementations/react/src/atomic/Stepper/src/stepper.types.ts
@@ -1,0 +1,9 @@
+import { HTMLAttributes, ReactElement } from "react";
+import { Step } from "./components/Step";
+
+export interface StepperProperties {
+  direction?: "vertical" | "horizontal";
+  children: ReactElement<typeof Step> | ReactElement<typeof Step>[];
+}
+export type StepperProps = StepperProperties &
+  Omit<HTMLAttributes<HTMLDivElement>, "children">;