+++ packages/styles/src/packages/atomic/stepper/index.ts
@@ -1,0 +1,5 @@
+import { styles } from "./rarui-stepper.css";
+
+export const stepper = { classNames: { ...styles } };
+
+export type { StepperVariants } from "./rarui-stepper.types";