+++ packages/styles/src/packages/atomic/stepper/rarui-stepper.types.ts
@@ -1,0 +1,4 @@
+import { RecipeVariants } from "@vanilla-extract/recipes";
+import { styles } from "./rarui-stepper.css";
+
+export type StepperVariants = RecipeVariants<typeof styles.stepper>;